move cache module under core namespace
